+++ XsltEditor/App.xaml.cs
@@ -8,8 +8,8 @@ using XsltEditor.Services.Interfaces;
 using XsltEditor.ViewModels;
 using XsltEditor.ViewModels.Dialogs;
 using XsltEditor.ViewModels.Main;
+using XsltEditor.Views.Windows;
 using XsltEditor.Views.Windows.Dialogs;
-using XsltEditor.Views.Windows.Main;
 
 namespace XsltEditor;
