+++ XsltEditor/ViewModels/Dialogs/GoToLineViewModel.cs
@@ -2,9 +2,8 @@ using System.Windows.Input;
 
 using XsltEditor.Infrastructure;
 using XsltEditor.Models.Base;
-using XsltEditor.ViewModels.Main;
 
-namespace XsltEditor.ViewModels;
+namespace XsltEditor.ViewModels.Dialogs;
 
 public class GoToLineViewModel : ObservableObject
 {
