+++ XsltEditor/ViewModels/SettingsViewModel.cs
