+++ XsltEditor/Views/Windows/MainView.xaml
@@ -1,4 +1,4 @@
-﻿<ct:Window x:Class="XsltEditor.Views.Windows.Main.MainView"
+﻿<ct:Window x:Class="XsltEditor.Views.Windows.MainView"
            xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
            xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
