+++ XsltEditor/Views/Windows/SettingsView.xaml
@@ -1,4 +1,4 @@
-<ct:Window x:Class="XsltEditor.Views.Windows.Main.SettingsView"
+<ct:Window x:Class="XsltEditor.Views.Windows.SettingsView"
            xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
            xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
